refactor(blockstore/engine): split engine.go 1213-line god file into 6 files
Pure file reorganization, zero logic change. Every function moved verbatim
with same body, signatures, imports.

- engine.go (433): Config, BlockStore, BlockStoreStats types; New, Start,
  Close, loadByHash; LocalForTest, RemoteForTesting, ListFiles, EvictLocal,
  EvictReadBuffer test helpers.
- readwrite.go (311): ReadAt, GetSize, Exists, WriteAt, Truncate, Delete,
  CopyPayload.
- flush.go (101): Flush, DrainAllUploads.
- stats.go (109): Stats, GetStats, populateBlockCounts, LocalStats.
- health.go (81): HealthCheck, Healthcheck, HasRemoteStore,
  SetRetentionPolicy, SetEvictionEnabled.
- read_internal.go (224): readAtInternal, ensureAndReadFromLocal,
  readLocalByHash, findRowCoveringOffset + rowWithOffset, blockRefHashes,
  computeFileSize.

v1.0 area #1 audit B-B (F-01: engine.go god file).
