+++ pkg/blockstore/engine/health.go
@@ -1,0 +1,81 @@
+package engine
+
+import (
+	"context"
+	"time"
+
+	"github.com/marmos91/dittofs/pkg/blockstore"
+	"github.com/marmos91/dittofs/pkg/health"
+)
+
+// HealthCheck verifies the store is operational by checking the syncer health
+// (which in turn checks the remote store).
+//
+// Legacy error-returning probe. New callers should prefer Healthcheck
+// (lowercase 'c') which returns a structured [health.Report] derived
+// from both the local and remote stores and satisfies [health.Checker].
+func (bs *BlockStore) HealthCheck(ctx context.Context) error {
+	return bs.syncer.HealthCheck(ctx)
+}
+
+// Healthcheck returns the engine's overall health, computed as the
+// worst-of of its underlying local and remote stores. The result
+// satisfies [health.Checker] so the API layer can wrap the engine in
+// a [health.CachedChecker] for /status routes.
+//
+// Derivation rules (worst-of)
+//
+//   - If the local store reports unhealthy → engine is unhealthy
+//     (we can't even serve cached blocks).
+//   - If a remote store is configured and reports unhealthy → engine
+//     is degraded (local reads still work, but new uploads will queue
+//     and the system is operating in offline-write mode).
+//   - Otherwise → healthy.
+//
+// The combined message preserves the worst-status component's message
+// so operators can see exactly which subsystem is at fault.
+func (bs *BlockStore) Healthcheck(ctx context.Context) health.Report {
+	start := time.Now()
+
+	if err := ctx.Err(); err != nil {
+		return health.NewUnknownReport(err.Error(), time.Since(start))
+	}
+
+	localRep := bs.local.Healthcheck(ctx)
+	if localRep.Status == health.StatusUnhealthy {
+		return health.NewUnhealthyReport("local: "+localRep.Message, time.Since(start))
+	}
+
+	if bs.remote != nil {
+		remoteRep := bs.remote.Healthcheck(ctx)
+		if remoteRep.Status == health.StatusUnhealthy {
+			// Local works, remote is unreachable: degraded — reads
+			// still served from local cache, writes will queue.
+			return health.Report{
+				Status:    health.StatusDegraded,
+				Message:   "remote unreachable: " + remoteRep.Message,
+				CheckedAt: time.Now().UTC(),
+				LatencyMs: time.Since(start).Milliseconds(),
+			}
+		}
+	}
+
+	return health.NewHealthyReport(time.Since(start))
+}
+
+// HasRemoteStore returns true if this BlockStore has a remote store configured.
+func (bs *BlockStore) HasRemoteStore() bool {
+	return bs.remote != nil
+}
+
+// SetRetentionPolicy updates the retention policy on the underlying local store.
+// Delegates to the local store's SetRetentionPolicy method.
+func (bs *BlockStore) SetRetentionPolicy(policy blockstore.RetentionPolicy, ttl time.Duration) {
+	bs.local.SetRetentionPolicy(policy, ttl)
+}
+
+// SetEvictionEnabled controls whether the local store can evict blocks to free disk space.
+// Delegates to the local store's SetEvictionEnabled method.
+func (bs *BlockStore) SetEvictionEnabled(enabled bool) {
+	bs.local.SetEvictionEnabled(enabled)
+}
